app/models_backend/api/requirements.txt
tensorflowcd
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy